fix(backfill): catch UnicodeDecodeError from validate_applied's spec read
Tier-3 CI-gate re-review (T10, round 9): the T9 fix's `except OSError`
around validate_applied() didn't cover malformed UTF-8 in spec.md --
read_text(encoding="utf-8") raises UnicodeDecodeError, a ValueError
subclass, so that failure mode still crashed main() uncaught with writes
left un-rolled-back. Widen the clause to (OSError, UnicodeDecodeError).

Split the rollback-test file's failure-safety tests (T9's two, T10's new
one, and the T8 symlink test) into a new
test_backfill_ac_provenance_cli_rollback_failsafe.py sibling -- the
original file crossed the 300-LOC guideline again after T10's addition.
